Tolerate frontend/rule without active services, add alb info to error page, calculate priority based on terms
http://jira.alauda.cn/browse/ACP-2015
http://jira.alauda.cn/browse/ACP-2052
http://jira.alauda.cn/browse/ACP-2051
http://jira.alauda.cn/browse/ACP-1091

+++ controller/alb2.go
@@ -43,7 +43,7 @@ func MergeNew(alb *m.AlaudaLoadBalancer) (*LoadBalancer, error) {
 		for _, arl := range aft.Rules {
 			rule := &Rule{
 				RuleID:          arl.Name,
-				Priority:        int64(arl.Priority),
+				Priority:        arl.Priority,
 				Type:            arl.Type,
 				Domain:          arl.Domain,
 				URL:             arl.URL,
